--- lib/hir-mir/src/validation.rs
@@ -1,1 +1,0 @@
-pub mod ffi_variant_type_validator;
